set ignored files

+++ broadband/src/main/java/com/tm/broadband/model/CustomerCallingRecordCallplus.java
@@ -20,7 +20,7 @@ public class CustomerCallingRecordCallplus implements Serializable {
 	
 	private Integer id;
 	private Date date;
-	private Double length;
+	private Integer length;
 	private String description;
 	private Double charged_fee;
 	private String original_number;
@@ -63,11 +63,11 @@ public class CustomerCallingRecordCallplus implements Serializable {
 		this.date = date;
 	}
 
-	public Double getLength() {
+	public Integer getLength() {
 		return length;
 	}
 
-	public void setLength(Double length) {
+	public void setLength(Integer length) {
 		this.length = length;
 	}
