+++ broadband/src/main/java/com/tm/broadband/util/CallingRecordUtility.java
@@ -116,7 +116,7 @@ public class CallingRecordUtility {
 								// Integer
 								if(arr.length > durationIndex){
 									if(arr[durationIndex] != null && !"".equals(arr[durationIndex].trim())){
-										ccr.setDuration(Double.parseDouble(arr[durationIndex]));
+										ccr.setDuration(Integer.parseInt(arr[durationIndex]));
 									}
 									
 									// String
